from upstream

+++ tests/ImageSharp.Tests/Formats/Jpg/Utils/LibJpegTools.ComponentData.cs
@@ -9,7 +9,6 @@ using SixLabors.ImageSharp.Formats.Jpeg.Components;
 using SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder;
 using SixLabors.ImageSharp.Memory;
 using SixLabors.ImageSharp.PixelFormats;
-using SixLabors.Primitives;
 
 namespace SixLabors.ImageSharp.Tests.Formats.Jpg.Utils
 {
@@ -52,8 +51,8 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg.Utils
 
             internal void MakeBlock(short[] data, int y, int x)
             {
-                this.MinVal = Math.Min((short)this.MinVal, data.Min());
-                this.MaxVal = Math.Max((short)this.MaxVal, data.Max());
+                this.MinVal = Math.Min(this.MinVal, data.Min());
+                this.MaxVal = Math.Max(this.MaxVal, data.Max());
                 this.SpectralBlocks[x, y] = new Block8x8(data);
             }
